Fully functional player HUD
Linked in to delegate event calls (I had to make them static to access). Prefab for HUD should work when dropped into a scene with the player. There is an alternate "hit" image for the player portrait that could be swapped to on damage. Figured there were higher priority tasks left